Poprotska/basket page

+++ src/scenes/dashboard/task/Task.jsx
@@ -5,6 +5,7 @@ import Delete from '@material-ui/icons/Delete';
 import Info from '@material-ui/icons/Info';
 import moment from 'moment';
 import Tooltip from '@material-ui/core/Tooltip';
+import { AlertDialog } from "../../../components/dialog/AlertDialog";
 import PopapAddTagToTask from './popapAddTagToTask/PopapAddTagToTask';
 import * as styled from './Task.styled';
 import Checkbox from '../../../components/checkbox/Checkbox';
@@ -23,6 +24,7 @@ class Task extends Component {
             newTaskName: props.nameTask,
             visiblePopapAddTagToTask: false,
             selectedTask: '',
+            visible: false,
         };
     }
 
@@ -37,6 +39,13 @@ class Task extends Component {
     showPopup = () => this.setState({ statePopup: true });
 
     closePopup = () => this.setState({ statePopup: false });
+
+    showAlertDialog = () => {
+        const { visible } = this.state;
+        this.setState({
+            visible: !visible,
+        })
+    };
 
     handleCompleteTask = (time) => {
         const {
@@ -117,7 +126,7 @@ class Task extends Component {
 
     render() {
         const {
-            display, statePopup, visiblePopapAddTagToTask, selectedTask,
+            display, statePopup, visiblePopapAddTagToTask, selectedTask, visible,
         } = this.state;
         const displayStyle = { display, color: 'rgba(0, 0, 0, 0.54)' };
         const {
@@ -233,10 +242,16 @@ class Task extends Component {
                             </styled.AddTag>,
                             <Delete
                                 aria-label="trash"
-                                onClick={() => actions.deleteTask({ idTask })}
+                                onClick={this.showAlertDialog}
                                 style={displayStyle}
                                 alt="Delete this task"
                             />,
+                            <AlertDialog
+                                visible={visible}
+                                onClose={this.showAlertDialog}
+                                value='Do you want to delete this task?'
+                                onConfirm={() => actions.deleteTask({ idTask })}
+                            />
                         ]) : null
                     }
                 </styled.Task>
